# vox_api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Vox API server",
        "title": "Vox AI SI API",
        "contact": {
            "name": "Bogdan Antonovich",
            "email": "programjibogdan@gmail.com"
        },
        "version": "1.0"
    },
    "host": "smarttranslator.store/api",
    "basePath": "/",
    "paths": {
...
        "/hub/{hub_id}/listen": {
            "get": {
                "description": "Subscribes the client to a hub's audio stream and delivers synthesized audio chunks in real-time via chunked transfer encoding. The stream ends when the client disconnects or the hub closes.",
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "hub"
                ],
                "summary": "Listen to audio stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hub ID",
                        "name": "hub_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio stream delivered as chunked audio/mpeg"
                    },
                    "404": {
                        "description": "Invalid hub ID (IsHubIDValid middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/hub/{hub_id}/publish": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Receives an audio stream, transcribes it via Deepgram, processes the transcription via Groq, and synthesizes speech via Fish Audio. All three operations run concurrently. Requires a valid user and hub in context.",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "hub"
                ],
                "summary": "Publish audio stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hub ID",
                        "name": "hub_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transcription language code (e.g. en, ru)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio stream processed successfully"
                    },
                    "401": {
                        "description": "Missing or invalid auth cookies (IsAuthorized middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid user ID or hub",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "415": {
                        "description": "Invalid content type (IsContentTypeValid middleware)",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.HttpErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.HttpErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        },
        "CookieAuthRefresh": {
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    }
}